Consolidate and document SQL AST shims (#22094)
## Which issue does this PR close?

- Related to https://github.com/apache/datafusion/pull/22069

## Rationale for this change

While upgrading sqlparser, it was not clear why there were feature gates
for the `sql` feature:
https://github.com/apache/datafusion/pull/22069/changes#r3204705488

There is a mode to avoid the `sqlparser` dependency, added by @timsaucer
in
- #17332

I think this feature makes sense but it is a little hard to understand
because it is implemented with a bunch of `#[cfg(not(feature = "sql"))]`
and there is no central place that explains the design.

## What changes are included in this PR?

1. Consolidate the SQL AST shim structures in a separate module
2. Document the design so it is easier to understand

## Are these changes tested?

Yes by CI

## Are there any user-facing changes?

<!--
If there are user-facing changes then we may require documentation to be
updated before approving the PR.
-->

<!--
If there are any breaking changes to public APIs, please add the `api
change` label.
-->
